+++ Controllers/PokemonController.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.Caching.Distributed;
 using Newtonsoft.Json;
 using pokelist.Core;
 using pokelist.Model;
+using pokelist.Services;
 using pokelist.Filters;
 namespace pokelist.Controllers
 {
